refactor: Inline parser
